requirements.txt
# Core ML
tensorflow>=2.13,<2.17
numpy>=1.24,<2.0
scikit-learn>=1.3

# Visualización
matplotlib>=3.7
seaborn>=0.12

# API
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.0
Pillow>=10.0
requests>=2.31

# Notebooks
jupyter>=1.0
ipykernel>=6.0
